# swagger.yaml
openapi: 3.0.3
info:
  title: Ibrat Lead Calling System API
  description: |
    Complete backend API for AI-powered lead calling system with admin and operator panels
    
    ## 🚀 Quick Start Guide
    
    ### **1. Authentication Flow**
    1. **Register User**: `POST /api/auth/register` (admin only)
       - Use test credentials: `john.doe@ibrat.com` / `ibrat123`
    2. **Login**: `POST /api/auth/login` with same credentials
       - Copy JWT token from response
    3. **Use APIs**: Add `Authorization: Bearer <token>` header
    
    ### **2. Test Credentials**
    - **Admin**: `john.doe@ibrat.com` / `ibrat123`
    - **Salesperson**: `jane.smith@ibrat.com` / `ibrat123`
    
    ### **3. API Access**
    - **Admin Panel**: Full system access
    - **Operator Panel**: Work interface only
    - **Real-time**: WebSocket connections
    
    ### **4. Swagger Features**
    - Click "Try it out" to test APIs
    - Use "Authorize" button for JWT tokens
    - View request/response schemas
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@ibrat.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.ibrat.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from /api/auth/login
        
        **How to use:**
        1. Register: `POST /api/auth/register`
        2. Login: `POST /api/auth/login` 
        3. Copy token from response
        4. Click "Authorize" button above
        5. Enter: `Bearer <your_token>`
        
        **Example:**
        ```
        Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjUwN2YxZjc3YmNmODZjZDc5OTQzOTAxMSIsIm5hbWUiOiJKb2huIERvZSIsImVtYWlsIjoiam9obi5kb2VAaWJyYXQuY29tIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNzM1NzE5NjAwLCJleHAiOjE3MzU3MjU2MDB9.example
        ```
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "+1234567890"
        role:
          type: string
          enum: [admin, salesperson]
          example: "salesperson"
        isActive:
          type: boolean
          example: true
        isAvailable:
          type: boolean
          example: true
        lastLogin:
          type: string
          format: date-time
        callStats:
          type: object
          properties:
            totalCalls:
              type: number
              example: 150
            successfulCalls:
              type: number
              example: 120
            totalDuration:
              type: number
              example: 7200
    Lead:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439012"
        phone:
          type: string
          example: "+1234567890"
        name:
          type: string
          example: "Lead Name"
        email:
          type: string
          format: email
          example: "lead@example.com"
        status:
          type: string
          enum: [pending, claimed, calling, answered, transferred, failed]
          example: "pending"
        priority:
          type: string
          enum: [urgent, high, medium, low]
          example: "high"
        callAttempts:
          type: number
          example: 0
        assignedTo:
          $ref: '#/components/schemas/User'
        tags:
          type: array
          items:
            type: string
          example: ["hot", "interested"]
...
    LeadCreateRequest:
      type: object
      required:
        - phone
        - name
      properties:
        phone:
          type: string
          example: "+1234567890"
        name:
          type: string
          example: "Lead Name"
        email:
          type: string
          format: email
          example: "lead@example.com"
        priority:
          type: string
          enum: [urgent, high, medium, low]
          example: "high"
        tags:
          type: array
          items:
            type: string
          example: ["hot", "interested"]
...
paths:
...
  /api/leads:
    get:
      tags: [Lead Management]
      summary: Get all leads
      description: Retrieve leads with filtering and pagination
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, claimed, calling, answered, transferred, failed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [urgent, high, medium, low]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Leads retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer

    post:
      tags: [Lead Management]
      summary: Create a new lead
      description: Create a single lead manually
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadCreateRequest'
      responses:
        '201':
          description: Lead created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Lead created successfully"
                  lead:
                    $ref: '#/components/schemas/Lead'
  /api/leads/upload-csv:
    post:
      tags: [Lead Management]
      summary: Upload leads via CSV
      description: Bulk upload leads from CSV file (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                csvFile:
                  type: string
                  format: binary
                  description: CSV file with lead data
      responses:
        '200':
          description: CSV upload completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "CSV upload completed"
                  summary:
                    type: object
                    properties:
                      totalProcessed:
                        type: integer
                      successCount:
                        type: integer
                      duplicateCount:
                        type: integer
                      errorCount:
                        type: integer
...